feat: Add Toaster component for displaying notifications
This commit adds the Toaster component to the project. The Toaster component is used to display notifications at the top center of the page. It accepts props for position and can be customized further. This addition improves the user experience by providing a visually appealing and user-friendly way to show important information.