feat: Add DeleteBook component for deleting books
This commit adds the DeleteBook component to the project. The DeleteBook component is responsible for rendering a form with a delete button for deleting books. When the delete button is clicked, the corresponding book is deleted from the database using the Supabase client. This addition improves the functionality of the application by allowing users to easily delete books.